requirements.txt
# AI & Threat Analytics Automation Testing Framework

# Core dependencies
pytest==7.0.0
pytest-html>=3.1.0
pytest-sugar>=0.9.7
pytest-emoji>=0.2.0
pytest-xdist==3.5.0  # Parallel testing
pytest-timeout==2.2.0
pytest-cov==4.1.0
allure-pytest==2.13.2  # Test reporting

# HTTP & API
requests==2.28.0


# ML & Data Science
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
sentence-transformers==2.2.2  # For embeddings

# NLP
nltk==3.8.1
spacy==3.7.2

# Data validation
jsonschema==4.20.0
pydantic==2.5.2

# Logging & Configuration
pyyaml==6.0
python-dotenv==1.0.0

# Security & Encryption
cryptography==41.0.7

# Testing utilities
faker==20.1.0  # Generate test data
hypothesis==6.92.1  # Property-based testing

# Performance testing
locust==2.19.1  # Load testing (optional)

# AWS SDK (if needed for cloud integration)
boto3==1.34.10
botocore==1.34.10

# Development dependencies
black==23.12.1  # Code formatting
flake8==7.0.0  # Linting
mypy==1.7.1  # Type checking
pre-commit==3.6.0  # Git hooks

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
